+++ packages/tokens/src/variables/space.ts
@@ -1,5 +1,5 @@
 export const space = {
-  1: '0.25rem',
+  1: '0.3rem',
   2: '0.5rem',
   3: '0.75rem',
   4: '1rem',