+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @haruo-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Add new component slider
+
+## 2.1.0
+
+### Minor Changes
+
+- Fix colors and components
+
 ## 2.0.0
 
 ### Major Changes