+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @haruo-ui/tokens
 
+## 4.0.0
+
+### Major Changes
+
+- Add new component slider
+
+## 3.1.0
+
+### Minor Changes
+
+- Fix colors and components
+
 ## 3.0.0
 
 ### Major Changes